requirements.txt
python-dotenv
pandas
numpy
tqdm
requests
pykrx
python-dateutil
beautifulsoup4

torch
torchvision
torchaudio
transformers
sentencepiece

scikit-learn
matplotlib
seaborn
